Flyttet valideringslogikk av jobber sentralt, samt implementerte files i skjema for tasks
