+++ database/validators/task.js
@@ -1,0 +1,7 @@
+/*
+  Import dependencies
+*/
+// const HTTPError = require('../../lib/vtfk-errors/httperror');
+
+// module.exports = validate = (job, task) => {
+// }
