+++ routes/v1/jobs.js
@@ -7,6 +7,7 @@ const Job = require('../../database/db').Job
 const Statistics = require('../../database/db').Statistic
 const dbTools = require('../../database/db.tools.js')
 const HTTPError = require('../../lib/vtfk-errors/httperror');
+const validateJob = require('../../database/validators/job')
 
 /*
   Routes
@@ -25,59 +26,8 @@ router.get('/', async (req, res, next) => {
 // POST
 router.post('/', async (req, res, next) => {
   try {
-    // Validate that the tasks dont have duplicated types
-    const job = req.body
-
-    const taskTypes = []          // Array for alle unike task typer som er funnet i jobben
-    const dependencyTags = []     // Array over alle unike
-    const dependencies = []
-    job.tasks.forEach((task) => {
-      // If the task should be processed by E18, but no data has been provided
-      if (job.e18 !== false && !task.data) throw new HTTPError(400, 'Data must be provided on all tasks for E18 to process them');
-      // Make dependencyTag checks
-      if (task.dependencyTag) {
-        // Task cannot have dependency tag when there is only a single task
-        if (job.tasks.length === 1) throw new HTTPError(400, 'DependencyTag cannot be set when there is only a single task');
-        // Task cannot be dependent on it self
-        if (task.dependencies && task.dependencies.includes(task.dependencyTag)) throw new HTTPError(400, 'Task cannot have it self as a dependency');
-        // Add the dependencyTag to the dependencyTagsArray if not already present
-        if (!dependencyTags.includes(task.dependencyTag)) dependencyTags.push(task.dependencyTag)
-      }
-      // Make dependencies check
-      if (task.dependencies) {
-        // If task has dependencies, but there are no other tasks
-        if (job.tasks.length === 1) throw new HTTPError(400, 'Task cannot have dependencies when there are no other tasks');
-        // Check if there are more dependencies than there are tasks
-        if (task.dependencies.length > job.tasks.length) throw new HTTPError(400, 'There cannot be more dependencies than there are tasks');
-        // Add any dependencies to the array that have not been previously added
-        task.dependencies.forEach((d) => {
-          if (!dependencies.includes(d)) dependencies.push(d);
-        })
-      }
-
-      // Add tasktype to taskTypes if not previously found
-      if (!taskTypes.includes(task.type)) taskTypes.push(task.type)
-      // else throw new HTTPError(400, 'task type cannot be added more than once: ' + task.type)
-    })
-
-    // If the number of unique dependency tags and dependencies don't match up something is wrong
-    if (dependencyTags.length !== dependencies.length) throw new HTTPError(400, 'There is a mismatch between dependencyTags and dependencies');
-
-    // Make sure that all the specified dependecytags has been used by the dependencies
-    if (dependencyTags.length > 0) {
-      dependencyTags.forEach((tag) => {
-        const match = dependencies.find((d) => dependencyTags.includes(d));
-        if (!match) throw new HTTPError(400, `The dependency tags ${tag} is set but not used`)
-      })
-    }
-
-    // Make sure that all dependencies has a correlating tag
-    if (dependencies.length > 0) {
-      dependencies.forEach((dependency) => {
-        const match = dependencyTags.find((t) => dependencies.includes(t));
-        if (!match) throw new HTTPError(400, `The dependency ${dependency} is used, but has not been set`)
-      })
-    }
+    // Validate and sanitize the job
+    const job = validateJob(req.body);
 
     // Create and return the job
     res.body = await Job.create(job)
@@ -117,13 +67,18 @@ router.get('/:id/tasks', async (req, res, next) => {
 router.post('/:id/tasks', async (req, res, next) => {
   try {
     // Find the job
-    const job = await Job.findById(req.params.id);
+    let job = await Job.findById(req.params.id);
     if (!job) throw new HTTPError(404, `The job with id ${req.params.id} was not found`);
     // Validation
     if (job.e18 !== false && !req.body.data) throw new HTTPError(400, 'Data must be provided on all tasks for E18 to process them');
 
     // Push and save the task
     job.tasks.push(req.body);
+
+    // Validate & sanitize
+    job = validateJob(job);
+
+    // Save changes
     const updated = await job.save();
 
     // Return the saved task
